Add `gsutil` based operation
Implementing all operations (`copy`, `sync` and verify) with google
cloud python package might not be as effective as using `rsync`
through `gsutil`.

So the support for `sync` and `verrify` operations are added via
`gsutil` cli and call to them through supprocesses. For this option to
work the `gsutil` must be installed and configured.

The new implementation based on `gsutil` is not tested yet.

Signed-off-by: Saeed Gholami Shahbandi <saeed.gh.sh@gmail.com>